+++ src/main/java/com/practice/problems/adityaverma/dynamicprog/knapsack/zeroone/parent/Knapsack01Memoized.java
@@ -1,6 +1,8 @@
-package com.practice.problems.adityaverma.dynamicprog.knapsack.zeroone;
+package com.practice.problems.adityaverma.dynamicprog.knapsack.zeroone.parent;
 
 import lombok.extern.log4j.Log4j2;
+
+import java.util.Arrays;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -17,18 +19,22 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  **/
 
 @Log4j2
-public class Knapsack01Recursive {
+public class Knapsack01Memoized {
+
+    private static int[][] cachedResult;
 
     public static void main(String[] args) {
         int[] weights = {1, 2, 3, 5};
         int[] values = {1, 6, 10, 16};
         int capacity = 7;
+        init(weights.length, capacity);
         log.info("Input:: Weights: {}\t# Values: {}\t# Knapsack Wt: {}", weights, values, capacity);
         int total = solveKnapsack(weights, values, capacity);
         log.info("Output:: Total knapsack: {}", total);
         assertTrue(total == 22);
 
         capacity = 6;
+        init(weights.length, capacity);
         log.info("Input:: Weights: {}\t# Values: {}\t# Knapsack Wt: {}", weights, values, capacity);
         total = solveKnapsack(weights, values, capacity);
         log.info("Output:: Total knapsack: {}", total);
@@ -39,24 +45,37 @@ public class Knapsack01Recursive {
         return knapsack(weights, values, capacity, weights.length - 1);
     }
 
-    public static int knapsack(int[] weights, int[] values, int capacity, int index) {
-        if (index == -1 || capacity == 0) {    // IF >> index out of bound OR knapsack capacity ZERO
-            return 0;
+    private static void init(int index, int capacity) {
+        cachedResult = new int[index + 1][capacity + 1];
+        for (int i = 0; i <= index; i++) {
+            int tempCapacity = capacity;
+            while (tempCapacity >= 0) cachedResult[i][tempCapacity--] = -1;
         }
+        System.out.println(Arrays.deepToString(cachedResult));
+    }
+
+    private static int knapsack(int[] weights, int[] values, int capacity, int index) {
+        // IF >> index out of bound OR knapsack capacity ZERO
+        if (index == -1 || capacity == 0) return 0;
+
+        // IF >> Result is cached then return
+        if (cachedResult[index][capacity] != -1) return cachedResult[index][capacity];
 
         // Current Value and Weight
         int prevIndex = index - 1;
-        int currValue = values[index];
-        int currWeight = weights[index];
+        int indexValue = values[index];
+        int indexWeight = weights[index];
 
         // IF >> current weight goes beyond knapsack capacity it EXCLUDED
         int prevResult = knapsack(weights, values, capacity, prevIndex);
-        if (currWeight > capacity) {
-            return prevResult;
+        if (indexWeight > capacity) {
+            cachedResult[index][capacity] = prevResult;
         } else {
             // ELSE >> current weight is within knapsack capacity it is INCLUDED
-            int include = currValue + knapsack(weights, values, capacity - currWeight, index - 1);
-            return Math.max(prevResult, include);
+            int include = indexValue + knapsack(weights, values, capacity - indexWeight, prevIndex);
+            cachedResult[index][capacity] = Math.max(prevResult, include);
         }
+
+        return cachedResult[index][capacity];
     }
 }
